+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - Add SimpleNumericalPulse to the Qblox pulse library. (!1074)
   - Fix `StitchedPulseBuilder` not taking its `t0` into account for VoltageOffsets. (!1089)
   - Add stack_pulses compilation node to the Qblox backend. (!1084)
+  - Add crosstalk_compensation to the Qblox backend. (!1117)
 - Compilation
   - Add internal logic to calculate pulse compensations. (!1070)
   - Refactor parts of the NCO/LO frequency determining logic, and remove the `parent` attribute from `InstrumentCompiler` and all subclasses. (!1069)